概率装配完成测试

+++ big-market-domain/src/main/java/cn/bugstack/domain/strategy/repository/IStrategyRepository.java
@@ -15,7 +15,7 @@ import java.util.List;
 public interface IStrategyRepository {
     List<StrategyAwardEntity> queryStrategyAwardList(Long strategyId);
 
-    void storeStoreStrategyAwardSearchRateTables(Long strategyId, BigDecimal rateRange, HashMap<Integer, Integer> shuffleStrategyAwardSearchRateTables);
+    void storeStoreStrategyAwardSearchRateTables(Long strategyId, Integer rateRange, HashMap<Integer, Integer> shuffleStrategyAwardSearchRateTables);
 
 
     int getRateRange(Long strategyId);
